chore(internal/protoveneer): support Status
Support the Status type from
google.golang.org/genproto/googleapis/rpc/status.

We do not attempt to convert it to a more friendly type at this time.
First, although status.Status is a proto message, it does have two
fields, Code and Message, that most users should find it easy to
work with.

Second, if we made a new type, we'd have to decide where to put it.
It will be user-visible so the support package isn't the right place,
and it's a common type that occurs in many APIs, so adding it to the
current package being generated is also wrong. We'd need to make
a new package for well-known veneer types that correspond to
well-known proto types.

Third, the other field of Status is of type `[]*anypb.Any`. I have no
idea how to handle an anypb.Any in a nice way, because it can refer
to any proto message, using a specially encoded string to refer to the
type. Best to not touch it until we need to, or ever.