feat(aiplatform/apiv1beta1): add methods to info.go
Add SetGoogleClientInfo methods for each client in the GAPIC.

Also, move the script that generates the file somewhere safe
(internal, for now). In #10272 we put it in the same directory
as the GAPIC, and an automatic process deleted it.

This means it's not really usable with `go generate`, because its
relative location varies by package, so remove that functionality.