chore(internal/protoveneer): use nil timestamp for zero time
Convert a zero time.Time to a nil timestamp proto, instead
of a non-nil proto representing the zero time.

This can matter to services that care whether a timestamp
field is present or not. For example, the UpdateCachedContent
RPC of vertex/genai behaves differently depending on whether
an update time is passed to it.