fix(storage): allow empty soft delete on Create
Currently setting an empty soft delete policy to disable the
feature does not work on bucket creation (only update). This fixes
the issue by forcing the library to send a zero value in this
case (sending a null does not seem to work).

Also adds this case to the relevant integration test.

Fixes #10380