feat(storage): read stall timeout
Adds a RetryOption to set a timeout on stalled reads during downloads.
If individual calls to body.Read() exceed this timeout, the connection
will be closed and reopened from the last read offset.

This timeout will also be applied to gRPC reads and to the initial
read request in future PRs.