fix(auth): make default client creation more lenient
Before this commit, `CloneDefaultClient` would panic if `http.DefaultTransport`
was set to something that's not an `*http.Transport`. It is very much possible
that this is the case if users assign a different `http.RoundTripper` to
`http.DefaultTransport`, for example for tracing purposes.

This commit renames `CloneDefaultClient` to `DefaultClient`, and changes its
behavior to check if the `http.DefaultTransport` can be cloned, and uses the
clone if possible. If it can't be cloned, it'll use the `http.DefaultTransport`
directly.