test(auth): use the same now when checking tokenstate
I am not sure if this will fix the test, but it is a better
behaviour either way I think. If this happens again my only guess
is it has to do with rounding depending on the initial time
captured. If this test fails again we should use a reletive offset
time instead of a fully random time.

Releated: #10950