fix(auth/oauth2adapt): convert token metadata where possible
This was missed in the orginal implmentation. The code is still not
prefect as converting from an oauth2.Token to auth.Token is lossy
because there is no way to extract out the whole metadata context.
Converting from auth.Token -> oauth2.Token is not lossy though.

This bug was discovered because some direct path checks look for
theses metadata values. Just to be safe we will preserve both forms
of these values in the final token metadata, if present.

Internal Bug: 376281562