feat(spanner): add ResetForRetry method for stmt-based transactions (#10956)
* feat(spanner): add ResetForRetry method for stmt-based transactions

Read/write transactions that are aborted should preferably be retried using the
same session as the original attempt. For this, statement-based transactions
should have a ResetForRetry function. This was missing in the Go client library.

This change adds this method, and re-uses the session when possible. If the
aborted error happens during the Commit RPC, the session handle was already
cleaned up by the original implementation. We will not change that now, as
that could lead to breakage in existing code that depends on this. When
the Go client is switched to multiplexed sessions for read/write transactions,
then this implementation should be re-visited, and it should be made sure that
ResetForRetry optimizes the retry attempt for an actual retry.

Updates https://github.com/googleapis/go-sql-spanner/issues/300

* fix: only allow resetting if tx is really aborted

---------

Co-authored-by: Sri Harsha CH <57220027+harshachinta@users.noreply.github.com>