feat(compute/metadata): add debug logging
The compute/metadata module is used by too many other projects today
so we should not force the dependeny on gax from this specific module.
To tie into our client library logging we have the callers set up the
integration. Because of this decision we have to vendor in some of the
helper functions from gax into this project as well.