feat(spanner): add option for LastStatement in transaction
Adds an option to indicate that a statement is the last statement in a transaction.
This allows Spanner to skip some validations during the execution of the statement,
and instead rely on the validations during the Commit.

This option will also be used by the database/sql driver for statements that are
executed directly on a connection instead of in a transaction.

This change also fixes a number of test cases so that the test time is reduced
from 2 minutes to approx 5 seconds.