fix(storage): add a safety check for readhandle (#11549)
By running some tests observed that default ReadObjectSpec looks like: 
`read_object_spec:{bucket:"projects/_/buckets/bucket-name"  object:"obj-name"  read_handle:{}}`

Which is not correct as we pass nil read_handle. Added an check in NewRangeReader and NewMultiRangeDownloader to only pass read_handle in read_object_spec when it is not nil.

Reason: Default value of readHandle that we attach to objectHandle will be []( same as default value of byte array).

In case nil read_handle is passed by user we would already be doing the slow auth hence adding the check should not be a issue in that case too.