feat(metadata): allow canceling GCE detection
this adds a new `OnGCEWithContext` function that allows passing in
a cancellable context when performing runtime GCE platform detection.
`OnGCE` is marked for deprecation, similar to the other accessor
methods in this package.

currently there is a possibility that during GCE platform detection,
the metadata address and the DNS resolution steps could simultaneously
block, which prevents the function from returning, or can delay its
return significantly. enabling the caller to pass in a context mitigates
this as the caller can specify a deadline on the context or otherwise
manage its cancellation.