fix(auth): allow non-default SA credentials for DP (#11828)
This effectively reverts #11255 with the slight difference in that I refactored out the compute engine token provider check as an individual func, which I will utilize in pending #11812. Because the bound token credentials should only be configured if users didn't provide credentials explicitly or the given one is compute engine type.

Fixes #11821.