add test for nested context deadline
if system info indicates that the machine is on GCE, we will allow
for the probes to take longer to complete in the case that the first
one fails, to account for flaky probes. this adds a test for that case.

in order to test this code, we had to make a few changes. first, the
client now exposes OnGCEWithContext--this is so that we can override
the inner HTTP transport to, e.g., replace it with a transport that
has an artificial delay. second, we had to enable overriding the
"does this host look like it's on GCE" check in a test execution
environment.