feat(auth): configure DirectPath bound credentials from AllowedHardBoundTokens (#11665)
If the `AllowedHardBoundTokens` contains "ALTS", we will create an `altsCreds` that's configured to fetch DirectPath bound tokens.

gRPC channels will use the bound token in ALTS connections and fall back to normal tokens in TLS connections.

Although `altsCreds` is created before we check `isTokenProviderDirectPathCompatible`, if `creds` itself is not a compute engine credentials, DirectPath will not be attempted in the end. So there is no risk of creating an incompatible `altsCreds`. Moreover, it will not fail if normal tokens continue to be sent in ALTS channels.

FYI: @yamandabbagh I have started working on the e2e testing so I'm opening this myself.

Like the "MTLS_S2A" case, there seems no good way to unit test this. I will verify the end-to-end flow before marking this ready for review.