feat(storage): takeover appendable object (#11977)
Adds a new method `NewWriterFromAppendableObject` which allows the caller to start writing to an appendable-but-unfinalized object in GCS.

Conformance/emulator tests waiting on full support in the testbench for takeover. Integration tests pass locally, but cannot run in kokoro since zonal buckets are not yet available in the region.

Also fixes a bug where Writer.Attrs is not set appropriately for unfinalized uploads in some cases.