chore(internal/postprocessor): don't format pb.go files
Today we currently run goimports in our post processor over all
files in the repo. Because of this the purposed changes from owlbot
always has a large diff from what is landed in our repo. This in
turn causes a lot of churn in the owlbot PRs as well as everyone
gettings added as reviewers onto the PR. With this change we will
no longer format the files that we don't contorl the generation of
-- the raw proto gen. The files are still valid go code and the
only difference is that imports won't be sorted in the style goimports
likes. The upside is it fixes the other issues outlined above.