feat(spanner): add option for how to call BeginTransaction
Adds an option to the Spanner client for how BeginTransaction should be called.

The default for read/write transactions that are executed in a transaction runner, is to use
inlined-begin. That is; the BeginTransaction is inlined with the first statement in the transaction.
This reduces the number of round-trips to Spanner by 1.

The default for statement-based read/write transactions and read-only transactions is to use an
explicit BeginTransaction RPC invocation. This requires one additional round-trip to Spanner, but
can be more efficient if the application wants to execute multiple queries in parallel as the first
statements in the transaction.

This change introduces an explicit option that can be used to change the above defaults for a
single transaction. This is especially useful for statement-based read/write transactions, as these
are used by the database/sql driver, and currently requires an additional round-trip for every
transaction.