feat(spanner): wrap proto mutation (#12497)
Add an API for wrapping an existing protobuf mutation in a Spanner mutation. This allows
applications or libraries that use protobuf mutations to use these directly with the
Spanner client library, without first having to translate these to a spanner.Mutation
struct, and then having the client library convert it back to a protobuf definition.