chore(internal/postprocessor): make module scanning config-based
This PR is a small refactor to the postprocessor, and makes some
internal implementation details configurable via the common config.yaml
file.

Doing this as we've recently been defining new individually-released
modules (bigquery and pubsub), and thus we've needed to modify and
re-release the postprocessor itself, not just update configs.