chore(internal/postprocessor): address logic inversion (#12595)
In https://github.com/googleapis/google-cloud-go/pull/12541 the config
for modules to skip was transitioned from being hardcoded into the
implementation to a config file.  In the course of that refactor the
detect logic was inverted, so instead of a skiplist the new config
mechanism acted as an allowlist.

This PR corrects that issue, and adds another test to assert the
property that skip paths don't appear in the detected module list.

To fully address this issue, we'll also need to bump the owlbot lock
once this PR is submitted.