fix(spanner): enforce only one resource header
Only one resource header (google-cloud-resource-prefix) should be
included in requests that are sent to Spanner. If the context that
is used for a request already contains a resource header, then that
header value will be removed, and only the resource header that is
set by the Spanner client will be included in the request.