refactor(storage): Pipeline gRPC writes. (#12422)
* refactor(storage): Pipeline gRPC writes.

Modify the gRPC writer to send additional data while waiting for the
current chunk to flush. This is a substantial refactor.

Per the Go io.Writer interface contract, we must never modify or retain
the slice that the caller provides to Write. However, that doesn't mean
we have to copy every byte into a writer-controlled buffer: we can refer
to the byte slice in place. Therefore, if callers call Write() with more
bytes than the chunk size, we can send them to the service immediately
as long as we don't return from Write() until we no longer need the
caller's slice.

This is an in-place refactor which retains the existing flush behavior,
so its benefits are most obvious in somewhat unlikely scenarios. For
example: a 100MiB upload to a bucket in a remote region with chunk size
256KiB can see a latency reduction of ~35%.

There are two followup investigations made possible by this refactor.
The first is to flush less frequently when the caller provides write
slices much larger than the chunk size. This may provide an even larger
throughput improvement for cases like the one above, and is
straightforward to implement.

The second is to flush more frequently when the caller provides write
slices much smaller than the chunk size. (e.g. split a 16MiB chunk into
2x8MiB sub-chunks, and flush each when they're full.) This can avoid
pipeline stalls in more scenarios, by increasing the likelihood that
part of the chunk is available to buffer data without waiting for a
flush acknowledgement inline.

* fixup! refactor(storage): Pipeline gRPC writes.

* fixup! refactor(storage): Pipeline gRPC writes.

Review comments

* fixup! refactor(storage): Pipeline gRPC writes.

fix typo

* fixup! refactor(storage): Pipeline gRPC writes.

Don't retain user write buffers after returning from Write

* fixup! refactor(storage): Pipeline gRPC writes.

Correct issues with reduced ack tracking.

* fixup! refactor(storage): Pipeline gRPC writes.

* fixup! refactor(storage): Pipeline gRPC writes.

Update attr size during completions