chore(bigquery/v2): remove release-please config temporarily
This PR removes the release-please config and manifest entry for
bigquery/v2.  Due to issues with how release-please constructs branches
we can only have one bigquery release PR open at a time, and the
bigquery/v2 release is effectively blocking releases for bigquery.

When we're ready to resume this, we'll revert these changes.