feat(bigquery/storage/managedwriter): allow overriding proto conversion mapping  (#12579)
Currently the usage of `bigquery.InferSchema` and `adapt. StorageSchemaToProto2Descriptor` methods to generate proto descriptors with tables that have fields of type `TIMESTAMP` can have compatibility issues, since the original struct has a `time.Time` field and in the proto descriptor it becomes an `INT64`. 

This PR adds an option to convert to Google's Timestamp Well Known Type (WKT), which is also accepted by the Storage Write API. I think we can't make it the default because some customer might be relying on unmarshalling JSON data with timestamps in the `INT64` format ( unix timestamp ) instead of a RFC3339 formatted timestamp string already.

Also we discussed adding options to the `StorageSchemaToProto2Descriptor` method before, on the improvement issue related to CDC helpers: https://togithub.com/googleapis/google-cloud-go/issues/10721

Naming is hard, but not sure what is the best name for the `WithTimestampWellKnownType` method. Open to suggestions.

Fixes #12569 
Supersedes #12578