fix: improve multi-path API handling
Some APIs have multiple paths with nested elements. For example, in
Spanner we have:

- google/spanner/v1
- google/spanner/adapter/v1
- google/spanner/admin/database/v1
- google/spanner/admin/instance/v1
- google/spanner/executor/v1

Subdirectories are created effectively relative to the module name,
both for clients and snippets. This commit fixes the version handling
for both clients and snippets, for the postprocessor.

This commit does not change the snippet metadata handling for the
release command; that will be handled in a separate commit
(refactoring much of this code into module.go).