fix(internal/godocfx): module detection when tidy errors
Sometimes package.Load detects some untidied modules missing .sum
enteries. In this case the first package being retunred in in the
form 'cloud.google.com/go/foo/...'. This is not a real package and
is present to only return some package level errors. We should not
try to determine the module from this entry as it is not present.

Fixes a nil pointer exeception.

Updates: b/441963312