feat: provide minimal onboarding for Librarian
This adds a state file with AI Platform, but does *not* make the other
one-off changes (e.g. reordering imports, removing the executable bit)
as those would then be undone by OwlBot.

The intention of this change is to allow the automation that creates
generation PRs to be tested, but we should not merge the resulting PRs.

Fixes https://github.com/googleapis/librarian/issues/1982