fix(internal/librariangen): remove go mod init/tidy from postprocessor (#12832)
For now, we're going to make the generate command not touch these at
all. The command doesn't have enough context to use the right
dependencies, so we should leave them as they are.

We therefore shouldn't have go.mod and go.sum in the state file in the
remove regexes.

See https://github.com/googleapis/librarian/issues/1921 for context.