fix(librariangen): improvements to release-init (#12842)
* fix(librariangen): improvements to release-init

This only creates files in /output when it actually needs to.
It also uses a variation of the snippet metadata logic present in
postprocess.go. We will want to refactor this further, but it will
require a common set of types for libraries, APIs etc.

* fix(librariangen): fix expected JSON for release init requests

(This reflects the CLI at the moment; see https://github.com/googleapis/librarian/issues/2035 for details - we may need to change this again.)