chore: refactor request/response handling and metadata updates
The build and generate commands have requests which are simply a
single library in each case, whereas release-init has a dedicated
request type containing an array of libraries. The configure command
will be similar to release-init.

By sharing the same definition of "a library" across commands, the
code to update snippet metadata becomes portable between generate and
release-init.

I've left the "request" package name as-is, but we could rename it
later if it becomes confusing.

One change in behavior compared with the previous code: we now throw
an error if the snippet metadata file doesn't exist, but doesn't
contain either a version placeholder or a version number. That would
indicate a problem with the snippet metadata file.