test(storage): fix append test race (#12858)
Checks to Writer.Attrs().Size before the Writer is closed have been in
some cases triggering the race detector post writer pipelining updates.

The Writer.Attrs() doc says that this shouldn't be called until after
Close() (see
https://pkg.go.dev/cloud.google.com/go/storage#Writer.Attrs), so I think
it is fine to just remove these checks.

This fixes #12785 for me locally, but it may still fail in the CI due to
issues in us-west4.