feat(internal/stategen): populate the latest googleapis commit
An alternative would be to pass this on the command line, but as we
almost always want the latest hash anyway, it makes sense to do it
automatically.

We could have used a GitHub library for this, but there's so little to
do, it felt simpler to just make the HTTP request directly.