fix: use a fresh context for rollback
Use a fresh context for the Rollback of a transaction that returned an error.
This prevents the Rollback from being cancelled / skipped if the context has
been cancelled or the deadline has been exceeded. Skipping the Rollback would
otherwise mean that the transaction holds on to any locks that it has taken
for longer than necessary.