fix(spanner): use fresh context for rollback (#12897)
Use a fresh context without a timeout for the Rollback call when a
transaction has failed. This prevents the Rollback from being skipped or
cancelled if the original context's deadline has been exceeded. This
again would lead to the locks that had been taken by the transaction to
remain in place until Spanner releases these after 10 seconds.