chore: migrate auth module to Librarian (#12974)
Towards https://github.com/googleapis/librarian/issues/2417

The reformatting of state.yaml is a one-time thing (and will happen with
whatever module is next migrated).

Example release PR: https://github.com/jskeet/google-cloud-go/pull/15
(to be closed when this PR is merged - that's just what would happen
next time there's a release).

@shollyman @quartzmo @codyoss I'd quite like to get *this* PR in quite
early, as there are one-off changes (such as introducing config.yaml,
and the reformatting) which will apply to every migration PR.

Once this is in, it would be good to pair with someone on the Go team to
do a release. (Run Librarian, create the release PR, merge
it, wait for the release to go out.) I'm not sure whether the changes to
the auth library are ready to release, or whether there's more to come
though.