fix(internal/librariangen): handle multi-versioned modules and root module (#13013)
If a library ID includes the version (e.g. "pubsub/v2") then we assume
that's the module root, and we don't move generated code into just
"pubsub".

Related to https://github.com/googleapis/librarian/issues/2305

(Once we've figured out how to handle all these situations, we
definitely need to document "how to create a new major version" etc.)

Additionally (second commit) a library with an ID of "root-module" is
deemed to be at the repo root even if it doesn't have a "." source path.
This is to be used for the root module in google-cloud-go, as per
https://github.com/googleapis/google-cloud-go/pull/13025