fix(storage): persist resource in append retry
In one retry case for appendable writes, the client receives
both an object resource w/ write handle as well as an error from
the server. After reconnect the new stream would not keep the
original object resource. This change ensures that the object
is not dropped.

I need to make a change to the emulator as well so that the
testbench behavior more closely mocks the service.