feat(bigtable): Add experimental Bigtable connection pool with custom load balancing strategy.  (#12882)
Configurable Load Balancing: The pool accepts a btopt.LoadBalancingStrategy on creation, which determines how it selects a connection for a new request. The following strategies are implemented:

LeastInFlight: Scans all connections and picks the one with the absolute lowest number of active requests.

PowerOfTwoLeastInFlight: Picks two random connections and routes the request to the one with the lower load. This is a common and efficient alternative to a full scan.

RoundRobin: The default strategy, which cycles through connections sequentially.